Added Main EXE file. (The Main Application)
EXE file was made using Pyinstaller and should hold all of the methods used in the project without further processing or library downloads.

There is no installation if it, This is a simple app that starts a terminal with various features.

When you open the app, you'll be greeted with 4 options, Start, Settings, Save, and Quit. You can access them by inputting the number selection. When in the Menu Systems, you must use the input prompt to quit. When in the Gamepad mode, you can either hit ESC or close the application. (It's worth mentioning that this program doesn't even save your hotkey-binds if you do not manually use the save function on the main menu.)

At no point should this app be asking for firewall and or security exceptions.  IF you find that you have a copy from somewhere else and it has such a problem, please discard it immediately.